requirements.txt
django>=2.1.0, <2.2.0
djangorestframework==3.7.3